Move liveness
